% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{fsrs_initial_state}
\alias{fsrs_initial_state}
\title{Initial memory state for new card}
\usage{
fsrs_initial_state(rating)
}
\arguments{
\item{rating}{Rating: 1=Again, 2=Hard, 3=Good, 4=Easy}
}
\description{
Initial memory state for new card
}
