% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{fsrs_next_interval}
\alias{fsrs_next_interval}
\title{Calculate next interval}
\usage{
fsrs_next_interval(stability, desired_retention)
}
\arguments{
\item{stability}{Current stability}

\item{desired_retention}{Target retention (0.0-1.0)}
}
\description{
Calculate next interval
}
