% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsrs_api.R
\name{Scheduler}
\alias{Scheduler}
\title{FSRS Scheduler}
\description{
R6 class for scheduling card reviews using FSRS algorithm
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Scheduler-new}{\code{Scheduler$new()}}
\item \href{#method-Scheduler-review_card}{\code{Scheduler$review_card()}}
\item \href{#method-Scheduler-get_card_retrievability}{\code{Scheduler$get_card_retrievability()}}
\item \href{#method-Scheduler-to_json}{\code{Scheduler$to_json()}}
\item \href{#method-Scheduler-print}{\code{Scheduler$print()}}
\item \href{#method-Scheduler-clone}{\code{Scheduler$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Scheduler-new"></a>}}
\if{latex}{\out{\hypertarget{method-Scheduler-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Scheduler$new(
  parameters = NULL,
  desired_retention = 0.9,
  maximum_interval = 36500L,
  enable_fuzzing = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Scheduler-review_card"></a>}}
\if{latex}{\out{\hypertarget{method-Scheduler-review_card}{}}}
\subsection{Method \code{review_card()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Scheduler$review_card(card, rating, review_datetime = Sys.time())}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Scheduler-get_card_retrievability"></a>}}
\if{latex}{\out{\hypertarget{method-Scheduler-get_card_retrievability}{}}}
\subsection{Method \code{get_card_retrievability()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Scheduler$get_card_retrievability(card)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Scheduler-to_json"></a>}}
\if{latex}{\out{\hypertarget{method-Scheduler-to_json}{}}}
\subsection{Method \code{to_json()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Scheduler$to_json()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Scheduler-print"></a>}}
\if{latex}{\out{\hypertarget{method-Scheduler-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Scheduler$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Scheduler-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Scheduler-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Scheduler$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
